1.2.0 Release
# Known Issues

- half8 "equals" and "not equals" operators don't conform to the IEEE 754 standard - Unity has not yet reacted to my bug-report in regards to their "half" implementation.

# Fixes

- Added preliminary safety cast to a float of the half value in toboolsafe() until Unity fixes their half '==' and '!=' operators according to IEEE 754

# Additions

### "quarter" precision floats and vectors

- "quarter" is an 8-bit IEEE 754 1.3.4.-3 floating point value, often called a "minifloat"
- It has a very limited range of [-15.5, 15.5] with an epsilon of 0.015625. All integers, aswell as i + 0.5, within that range can be represented as a quarter
- Type conversion from - and to quarters also conforms to the IEEE 754 standard. In detail, casting to a quarter performs rounding according to a) its' precision and b) whether or not the more precise value is closer to 0 or to quarter.Epsilon. NaN and +/- zero preservation, aswell as preservation/clamping to +/- infintiy was also implemented
- "==" and "!=" operators for vectors conforming to the IEEE 754 standard were implemented (unlike, currently, Unity's "half" type). All the other boolean- and arithmetic operators were implemented for the base type only, which will return single precision results (for arithmetic operations). For vectors, quarter vectors are to be (implicitly) cast to single precision vectors first, until/if Unity chnages their "half" implementation.
- Type conversions from - and to all other single value and vector types were implemented
- Full function implementation within the library was added, including: abs(), isnan(), isinf(), isfinite(), select(), as[s]byte/asquarter(), vrol/r(), vshl/r(), tobool[safe]() and toquarter[safe]()

### Fixed Oversights

- Added missing type conversions from - and to half8 for (s)byte8, (u)short8 and (u)int8 vectors
- Added missing type conversions from - and to half8 for booleans and boolean vectors

- Added half "select" functions

- Improved the performance of unsafe boolean-to-half/float/double functions

- added (preliminary?) "abs", "isnan", "isinf" and "isfinite" for half and half vectors, eliminating unnecessary casting

+++ Runtime/Functions/float8 Unity Mathematics API.cs
@@ -25,7 +25,7 @@ namespace MaxMath
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
         public static bool8 isnan(float8 x) 
         { 
-            return (asint(x) & (int)bitmask32(31)) > 0x7F80_0000; 
+            return (asint(x) & bitmask32(31)) > 0x7F80_0000; 
         }
 
         /// <summary>       Returns the componentwise reciprocal a float8 vector.       </summary>
